unsloth-nvidia/requirements.txt
# Engram PoC - Unsloth/NVIDIA Dependencies
#
# Install PyTorch with CUDA first:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
#
# Then install these dependencies:
#   pip install -r requirements.txt

# Unsloth - Fast LoRA fine-tuning (install from git for latest)
# pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
unsloth>=2024.8

# Core transformers stack
transformers>=4.43.0
datasets>=2.20.0
accelerate>=0.32.0
peft>=0.12.0

# Training utilities
trl>=0.9.0
bitsandbytes>=0.43.0

# Data processing
pyyaml>=6.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
